Bug: 51721
Fixed upgrade code so that we use a rename operation to move cache/upload contents over to upload directory instead of the previous copy/delete operation which resulted in timeouts for systems with many files.  Moved the code to repair the upgrade_history table to the end of the scripts (silentUpgrade_step2.php and end.php) to correctly rename the filename column value of cache/upload files that get moved to upload directory.