Bug #54124: Project Task/Activity subpanel, when inline create Meeting, see None vs 15 min
I fixed selectors for selects and inputs. Project Task has own duration, date_start & date_end fields because of it document.getElementById was returning them instead of Meeting fields.
Also that fix is not correct by my opinion. We should have ability to add prefix for id of elements in subpanel. Because right now we have duplication of id of elements on that page.