Bug #54230: Report name containing javascript causes PHP Notices as well as Browser errors to apepar
There was only really problem that report_name is not decoded while saving. Because of it strip_tags function was doing nothing on that string.
Right now if you try to save tag in report name - it will be removed. In our case our report will get title "untitled" because after strip_tags name is empty.