Bug #53650: Workflow Type Templates not saving Type upon upgrade to 6.5.0, usable as Email Templates
I've moved update query to end.php because that is only place where I can to query after as db structure was updated.
Also I can't be sure in condition about current_db_version because if someone already installed 650 on 644 and got that bug - after upgrade to 653 he will still have that problem.