Bug 54700 & Bug 55007 - Manager Chart Not Showing Adjusted Values
Makes the Manager chart display the adjusted values based on what's in the filter.  This also makes the Manager chart default to the included in forecast no matter what the filter is selected on the side.

Note the adjusted values will only show up for the included items.  If you switch the filter to show the pipeline, the uncommitted values will be the total from the ops while the committed value will be the adjusted value.