Adds fixes to make the enum fields work properly for filters.
This starts the redefinition of how the filters are set for other views, which still needs to be fully defined.  For now, it changes the object getting set on the selectedCategory attribute in context.forecasts from {id: <id>, label: <label>} to {<id> : <label>} so that it can have multiple filters set as required by "Project lolrus".