Bug 55172 - Cumulative total is not calculated properly in the manager's Pareto chart.
This is a refactor, so that if there is no committed forecast, it will default the amount in the manager worksheet to 0 and also make the chart data be 0 for best/worst/likely.

I also rewrote the ForecastsWorksheetManagerApiTest to be One Test, One Assertion, so I could test the change i just made.