Eliminated MailerConfig because not all mailer libraries are going to use the same configs. For now it seems easiest to just use an array like a dictionary to store the configs.
Renamed "from" to "sender" and made other improvements to Mailer. Added tests for Mailer, but more are to come.

Simplified MailerApi so it's now only concerned with sending mail per the default configs of Mailer, which are the default PHPMailer configs. It is ideal to get things working with one sending strategy at a time, since we'll need to support multiple strategies at least initially.