Fixes for unit tests & logic
VardefManager::addTemplate method populates $GLOBALS. Because of it we have to remove picture field from $GLOBALS['dictionary'] intend of local $dictionary. In some cases $dictionary is local variable in vardef.php.

For example RebuildRelationship.php calls SugarBean::createRelationshipMeta and pass to it empty array as $dictionary. SugarBean::createRelationshipMeta include vardef.php and we are getting that $dictionary in vardef.php is not $dictionary from global scope.