Bug #48482: in Reports, the multiselect fields do not filter correctly
^ symbol is added to queries. Also changed logic of "is not one of".
if we have leads with values "B", "B,BD", "BD,BDE" then if our filter is "is not one of" "B,BD" we get result as "B" and "BD,BDE".
Previously we were getting all 3 leads because there was OR in condition. If you need old behavior please replace "AND" to "OR" on the line 32.