Implement the backend for storing attachments to the activity stream.
Each attachment is stored as a separate note. Since we do not have
access to the value property of the file input tag, we upload the base64
representation of the data, and then decode it server-side. We then
store the file using the GUID of the note created, just as it is done
with real notes.