Bug54696 - Conversion Field in Mobile View
The currency format for mobile fields aren't well implemented. The list params defined in studio doesn't have all the information that is being passed to the sugar_currency_format smarty plugin, which is correct. The currency should be passed like it is implemented for the SugarFields/Currency/ListView.tpl

Furthermore, the mobile fields are all set on the smarty template, unable to make customizations. This should be refactored on a next major version.

Add phpUnit tests.