Bug: 56711
Changed the code in EmailUI.php to not be hard-coded to look for the custom editviewdefs.php file.  Added test to class EmailUITest.php to check for output against all six out-of-the-box modules.

The previous code somehow worked by using the old meta parser code that would attempt a best guess in creating the editviewdefs.php file.  How that worked fine and how the custom module hard-coded value crept in all this time is a mystery to me.