Bug56642 - Quick Create Contact or Lead - Why is email not auto populated
This bug is already here for too long.
All the code using this widget is expecting to use the id as the count (total) of SUGAR.EmailAddressWidget for each module.

The class usage of SUGAR.EmailAddressWidget.instances makes no sense, since we are using a singleton pattern and this has the instances for each module already created. (This was introduced on the "Merge branch '6_4_4' into caramel" - 68dab813ae99cd178af874faf831e3d89d28004d)