Bug: 56433
Update SaveOverload.php so that we keep the product values in sync with the opportunity changes.  This is needed now that we are deriving the worksheet data for the individual view off of products.  The original bug (56433) was fixed by changes to the click to edit code, but a new problem occurred in that the currency_id value of the opportunity was not synced to the product (thus the worksheet data would never reflect any new currency id changes).