SFA-222 story.
This fixes the problems with reassigning the Notifications module for a user.  The Notifications module does not have team_id and team_set_id columns.  The previous code was simply trying to filter out the EAPM module and was not a good enough check.  This fix updates reassignUserRecords to check if the team_id and team_set_id columns exist for the targeted tables before adding the code to update team_id and team_set_id.  Added SugarTestNotificationUtilities to allow for quickly creating and removing created Notification beans.  Added ReassignUserRecordsTest.php.