Bug: 56801
Changed Products vardefs.php file so that date_closed and date_closed_timestamp are not required.  This is because some other modules do not use Products as an extension of Opportunities as is the case with Forecasting.  Also, the best, likely and worst case values were incorrectly tagged as ENT only.