Bug 56641 - Added JSONP support
This small patch adds JSONP callback support when calling the REST interface from external JavaScript, ie. with jQuery:

    $.ajax({
        url: config.endpointURL + "?jsoncallback=?",
        data: { method: methodName, input_type: "JSON", response_type: "JSON", rest_data: JSON.stringify(params) },
        dataType: 'jsonp',
        success: callbackFunction,
        error: failFunction
    });

jQuery creates a random jsoncallback GET parameter as in the JSONP standard and the server wraps a JS function call around the JSON response. jQuery (and other libraries) then include the URL with a tag so that the random function gets called and receives the data. This is to avoid cross-domain origin issues.

This only works for GET requests. Numerous blog entries tell that POST requests to the REST API work through YQL as a work around.

With this patch I'm able to invoke the SugarCRM ReST API from an outside, local JavaScript.