Fixed a bug where the plain-text body was not being set in the case of a multi-part email. Refactored the code around how the message parts are set.
A hack was added to SmtpMailer::transferBody which guarantees that from_html will be called on each message part prior to looking for certain strings found in the HTML part. If from_html was not called, then some HTML characters would remain encoded and searches for the decoded strings would fail.