Fixes issue with default month not getting set in dropdown because default template uses value, instead of def.value.
Refactors the way we determine if this view should be read only.

Fixes the way the number of days in the month is calculated, which was getting set incorrectly since we were basing it off the current date, and only setting the month of the new date without resetting the day.  Thus, if the current month has 31 days, and the month we set it to only has 30, the Date object would get incorrectly set to the month following the selection.