Bug 55641 - CSV export gets cut off half-way for multibyte data
Fix `Content-Length` on header response to be in decimal value in OCTETs. Normally a strlen would be used.
Use `mb_strlen()` with the second param as `8bit` to fix possible definitions on the `php.ini` file for the `mbstring.func_overload`.
See http://php.net/manual/en/mbstring.overload.php