Fix for test
Because of fix #58100 some logic was changed. I restored it.
The main reason is that by default we should get tasks as VTODO. We should use correct variable with correct name like "tasAsVTODO". It allows us to detect should tasks be included in current method or be returned from CalendarActivity::get_activities.