Bug: 58154
Made changes so that the timeperiods created will take into account the current date and generate timeperiods from that point forward/backward even though the start month and start day may be different.  Also fixed so that the start/forward calculations are based off of the current leaf timeperiods instead of the setup start month and day as the median point.  Added include/SugarForecasting/Filter folder and class TimePeriodFilter for this to allow extension.  Added unit tests.