Bug: 58648
This is a bug with the uw_utils.php upgradeUWFiles method.  It was passing in the array of the already searched files into the findAllFiles function which would result in the additional entries already found being appended to the array.  The fix is in uw_utils.php, but since that is run before the upgrade, the changes will be picked up upgrading from 7.0.0 since we cannot make changes to 6.5.x and 6.6.x code base at this point for these issues.