Bug: 59409
Made updates to TimePeriod.php to split up creating TimePeriods for upgrades into a separate function.  This way it could be easier to unit test as well as to manage.  Added unit test support for upgrade of TimePeriods with support to create "backward" TimePeriods leading up to the selected start date while preserving the current TimePeriod's id.