Changed the data provider pattern for the recipients field's jasmine tests so that each case is run independently.
Fixed a broken test that was due to a typo. The email address should not have ended in ".come"; ".com" was correct. Finding this typo made aware the fact that the test was expecting the wrong return value. It was just passing previously by mistake because of the typo and a misunderstanding of the output.