Bug: 61208
Made changes to Product.php so that the create_new_list_query and create_export_query only retrieves Products where opportunity_id is NULL or opportunity_id = ''.  The latter check for '' is done because of the way SugarBean and the DBManager classes handle id type fields.  They are just set to blank if no value is provided rather than NULL.