Bug: 61291
Provide server side translation for fix.  The problem is that ServiceBase uses the preferred language of the user to determine the global current_language variable.  Client side translation would be a much larger refactor and would probably not be feasible for 6.7 at this point.  Unless the logic in ServiceBase is changed then this may be the route for now.