Bug: 61486
The recursive_queries capabilities was merged back into 6.7 accidentally.  This should be set to false.  MysqlManager cannot support recursive db queries.  The /Forecasts/reportees endpoint should fall back to use the Non-Recursive hierarchy code.  Sugar7 is supposed to drop support for MysqlManager so this should no longer be an issue in Sugar 7.