Need to return the Inbox ID if the configuration is an InboundEmail record; shouldn't always return the Config ID, which is actually the OutboundEmail record. This is because an OutboundEmail ID can be used by more than one InboundEmail, so it cannot be considered unique in the context of an Inbound configuration.
We also don't care about the name of the configuration since it's never used in client-side code. The Config Name actually refers to the name of the OutboundEmail record. When the time comes for introducing the configuration name to the API, we must decide whether the OutboundEmail or InboundEmail name should be returned in each context.

jira: MAR-548