The Mail API expects the recipients to be formatted as an array of json objects like {email: email, name: name}. When recipients were structured as '"Name" <email>' in the Email Compose UI, the email addresses weren't parsed. '"Name" <email>' was not a valid email address, so it was rejected. The email could not be sent and an error was returned to the user.
The recipients field maintains a collection of recipients in the format required because we ultimately want to use a real data structure to store the recipients in memory, as opposed to constructing and deconstructing a string over and over. However, we are forced to use a string at this time because it simplifies how we handle user input, which we've chosen to ignore for now. When the user clicks the send button, we now send the collection of recipients, instead of the string, to the server.

Additionally, we now close the drawer when an email is sent successfully.

jira: MAR-546