BUG N/A: - Added forecast admin settings to make sure the api test would work as expected within the suite - changing odd edge cases tests to utilize a fixed date to set now to, so that the dates can be reliable fixed points in time and testable.  Currently they were floating, so as time progresses the tests would fail because they're expectations would be off.
The most problematic was the testOddEdgeCases.  The dataprovider was attempting to create data surrounding the current date.  It was adjusting year at the beginning of the dataprovider, but not throughout all of the arrays of data being created.  When a date would pass some of the dates in the provider, the year wasn't adjusting again inside of that array, so it would be creating the wrong expectations.

Take this example: https://gist.github.com/jspriggs/b20cbed5154da794a6c2

The first 2 elements are used for forecasting setup to determine what kind of time periods are used.  Then the start and end dates are defined.  The start date is used for the setup, the end date is used as an expected value for an assertion.  The array at the end is a rundown of the current leaf timeperiods expected so we could further test that.  It's over reaching in my opinion looking back at it that it would be testing so may parts.

But for an explanation.  If currentYear is based off current date and today is March 29, 2013, your current year is 2013, and the data provider will make the expectation that the current timeperiod created will be from March 31, 2012 to March 30 2013.  A few days pass and today is April 1, 2013.  CurrentYear is still defined as 2013, but because the action of building the timeperiods was also readjusting the dates to determine what the "current" timeperiod would be, now the created timeperiod will be for March 31 2013 to March 30 2014.

By making the update in my PR, I made it so the timedate->now is set to a fixed point in time, so the "current" timeperiod can always be expected to fall within testable parameters without doing confusing date computations to force the test data to build around expected output.

The other fix was simply because the forecasting settings were being skewed by a previous test somewhere, but it's a better practice when dealing with testing the timeperiods to set the forecasting settings in the setupBeforeClass to make sure that they are within the expected parameters of the test environment in case there is existing data that could be skewing as well.