Fixes a bug where labels and badges that were not links were still acting like links because of the ".label:hover, .badge:hover" declaration. Now, only labels and badges that are links will behave like links and the base styles for labels and badges are consistent, regardless of whether or not the label/badge is a link.
The following are acceptable ways to create labels and badges (ranked in order from best to worst usage pattern):

<a class="label">Matches selector: a.label</a>
<span class="label"><a>Matches selector: .label > a</a></span>
<span class="label"><span><a>Matches selector: .label a</a></span></span>
<a><span class="label">Matches selector: a > .label</span></a>

The following is not an acceptable way to create a label or badge:

<a><span><span class="label">Matches selector: a .label</span></span></a>

"a .label" is very greedy, so it might lead to unintended consequences. For example, the scenario where the element with ".label" is nested far beneath the anchor tag.

Additionally, the font color for labels and badges is changed to white per a recommendation by Wes Moran.

jira: MAR-636