Elastic: refactor query builder to ignore unsupported types
We are already skipping datetimecombo fields into the fields to search
when building the elastic query. On current implementation we hit the
same issue on integer fields (ie case_number).

We should not exclude those fields based on the sugar type, but on the
actual elastic field type those get mapped to. If for example we define
in the vardefs a date fields as a string, this field becomes searchable.

This commit uses the SugarSearchEngineElasticMapping class which has all
the logic available to determine the elastic field type. The logic
inside getSearchFields() has been refactored to avoid duplicate logic to
handle (optional) boost values and prepare the final search field list.