Elastic: Clear SugarSearchEngineMetadata cache entries when clearing searchCache from Quick Repair and Rebuild. This also needs to be called when saving a field from Studio to make boost changes available directly on the elastic queries and enable/disable a field on the fly.
Note: when a field is enabled (by setting a boost), this still requires
a full reindex. However with this code change we can alter boost values
for elastic on the fly in studio as they are no longer maintained in the
index but are added on the actual query on a per field base.

Indepent of this "new" functionality, the cache entries created by
SugarSearchEngineMetadata are never purged so I wonder how changes are
actually propagated today. This commit has been tested on an APC enabled
web server.