BR-148 - Portal ACL's not applied in the FilterApi
I have added the methods needed to SupportPortalVisiblity to utilize SugarQuery and add the correct Visibility clauses to the queries.  This will allow the portal to use filter.  I have also fixed SugarQuery to correctly add a raw SQL to a where.

Added a test

Added a test and fixed the issues that came up during the test

Stas' suggested fixes for the unit test