BR-149 and BR-243 - Password fields should be blanked in the API
The API should not show password fields [type=password in vardefs].  It
should blank these out.  I have made the change to do this, and also
updated the SugarACLUsers to allow admins and self to set user_hash.

BR-149 and BR-243 - Password fields should be blanked in the API

The API should not show password fields [type=password in vardefs].  It
should blank these out.  I have made the change to do this, and also
updated the SugarACLUsers to allow admins and self to set user_hash.