[No RTC Task] Unit test(s): Unit tests for elastic.
This commit is for all unit tests for elastic core changes made by the sfa team, EXCEPT for
the unit test for SugarSearchEngineElastic::addFacets(). This unit test was left out deliberately
because the pull request that includes this new method as not been merged into nutmeg_ibm yet.

The pull request where addFacets() was added to SugarSearchEngineElastic is:
https://github.com/sugarcrm/Mango/pull/11857

A separate commit with the unit test for addFacets() will be committed later.

For elastic search.