Shares a record from sidecar modules (fixes SC-1211)
If the user doesn't have Email composer configured, use the mailto browser
feature to share the current record.
If the user has the Sugar internal email configured, use the Email composer
to prepare to send a message.
Both methods pre-populate data to subject and body. On Email composer body in
HTML is preferred.