jira: MAR-1073 Entering this test case (prior to PHPUnit_Framework_TestCase::setup being called), $GLOBALS['dictionary']['Activity'] is null. When instantiating Activity beans, SugarBean's $loaded_defs['Activity'] is an array of what appears to be the expected key-value pairs. This indicates that a previous test left the dictionary in a bad state (i.e., out of sync with $loaded_defs). Thus, this test case will fail because the expected vardefs cannot be loaded into $GLOBALS['dictionary']['Activity'] under normal operating procedure; and new Activity beans cannot be constructed as defined by the vardefs (e.g., the relationships are missing, which makes it impossible to add a related record to an Activity).
It was discovered that the following three tests that leave the dictionary out of sync. However, when running all unit tests with those three tests skipped, this test still fails in the same way. It seems that one or more other tests have the potential of leaving the dictionary out of sync, but only do so when the first three culprits are removed.

Bug53218Test::testWorkFlowConditionModules
SaveRelationshipChangesTest::testHandleRemainingRelateFields
Bug51596Test::testSearchAndUpdate

After much investigation, the pragmatic thing to do was to force the vardefs to be reloaded so that no outside forces could negatively impact this test. This fix is not ideal, but it is necessary for the time being.