[Meetings] [Core] This test is not related with Meetings but with SugarBean
See PR https://github.com/sugarcrm/Mango/pull/6597/files
The unit test should be made against SugarBean if we really want to
support failures of relationships with bad names (which we shouldn't).
If there is a bad `$rel_link` if should be fixed on the metadata or
vardefs, shouldn't be assumed that its ok to fallback gracefully.