Fix dashablelist for BWC modules.
We convert the old listviewdefs on the fly for now. This *must* be
removed by 7.2 (or as soon as we no longer have any record-based
modules that can be added as a dashable list), as we use a blocking
XHR request which causes the browser thread to lock up while it waits
for the legacy metadata response from the server.

Although the dashablelist dashlet will show the correct fields (and
take into consideration Studio fields as well), it will not properly
show special fields such as datetimes or teamsets if the old vardefs
did not explicitly mention this in their vardefs through the 'type'
key. One example of this would be the teams field on Calls. Currently,
it will show [object Object] when added to the dashlet. This is
because it falls back on the 'base' field.

jira: MAR-1318