Fix team info indexing for full-system indexes.
The old check when generating the query only checked to see if the
team fields (team_id and team_set_id) were set to be indexed by the
FTS engine. However, these fields must be indexed for all modules that
implement team security, regardless of whether the vardefs have those
fields set to be indexed or not. This is because we add a filter for
team_set_id when querying the FTS engine when querying as a non-admin
user.

jira: MAR-1323