SC-1892 - Change so today on planned activities uses equals
This change makes planned-activities dashlet use $equals instead of
$lte.

When sending in date only in a datetime field $equals will be amended to
be a between that will be your TZ's 00:00:00 and 23:59:59.

I have verified that the change works as expected in the dashlet and
that history is working as expected as well.