SFA-2290: fix currency unformatter to only return valid values
The currency unformatter blindly unformats values from the field.
If the field was invalid and the unformatter returns an invalid number,
the user could see unexpected results in the field and the validator could
also give unexpected results. Now the unformatter will leave a value alone if
it cannot unformat it to a valid numeric value.