Separate the module list layout from the menu view (SC-2360)
This provides us more flexibility to extend or change the menu per module
basis (like sugar cube = Home).
Improved performance on manipulation of the menu. It is now possible to work
with the menu without causing memory leaks.

The menu has two renders for the main nav bar: a short and a long version.
While the long version is able to provide the dropdowns of the actions etc,
the short is able to provide other customizations that we will have in a
recent future.