SFA-2332: fix api unformatting to not use server-side unformatting by default
With sugar 6, all field unformatting happened server-side. With sugar 7, fields unformat themselves
client-side then pass unformatted content server-side, often ending in double-unformatting.

The initial fix for this was to test server-side if the value is numeric, and if so don't unformat.
However there is an ambiguity where . and , swapped (1000s sep is dot) and something like "5.000" is
passed into the server-side unformatter. There is no way to distinguish this from a formatted or
unformatted value.

The proper approach is in SugarFieldBase. Currently we have save() and apiSave() which just
wraps save() and consequently calls unformatField(). Now apiSave() will call apiUnformatField()
which (by default) does nothing. Normally S7 fields will need no server-side unformatting, however
you can override apiUnformatField() if there is a need to do so.

apiUnformat() is deprecated and will continue to be used by the FilterApi endpoint until the next
API version to avoid breaking code that extends it.