MT-920: Use a higher curl timeout value for full indexer (async) and cleanup unnecessary timeout handling.
There is no need to have a separate fts_test_timeout apart from the timeout parameter which can already be configured. Also note that this so called "timeout" is *NOT* an actual connection timeout, but the maximum amount of time a curl execution is allowed. This parameter is passed in the Elastica backend and used as the value for CURLOPT_TIMEOUT (see ) and has nothing to do with an actual network connection timeout like for example

The full indexer timeout is hard coded and set to a high value as this execution happens asynchronously. In case the ES cluster is not able to consume the requests within this hard coded timeout, the batch sizes can be configured and should be lowered instead.