SFA-2369: remove currency symbol from quantity in quotes PDF report
This was a bug where both type currency and type decimal were treated as currencies.
The check for type decimal has been removed, as all db types have been corrected since 6.5.

This was already fixed in Bug 66536 but never pulled forward.  Instead of cherry-picking
this was fixed manually to clean up unecessary parenthesis as well.