Filtering by removed fields shouldn't break application (fixes SC-2551)
When filtering by a field that was removed on studio, we get a 422 from the
server. To avoid that we should check if the filter isn't going to be applied
on fields that don't exist.

Based on acceptance criteria defined in previous versions, we should also
remove fields that no longer exist in the search metadata.
See https://sugarcrm.atlassian.net/browse/SC-2551?focusedCommentId=126979&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-126979