SFA-2375: Product Groups Sorting on Opps
This was saving correctly with the proper indexs but when it was loaded and sorted, the products where not loading their product index.  This was due to a change on how SugarQuery handles the Relationship Role fields and BR-518, Once the Role field def was added to products/vardefs.php the sorting started working again and everything was coming up how it was saved on the quote.