[BR-1562] List view performance improvement
FilterApi uses DISTINCT clause when construction the list view query. This has a bad impact on performance and there is no good way in optimizing this properly for bigger data sets. Also the paging with or without DISTINCT is broken in some scenarios because php code is used to resolve distinct record ids anyway. Therefor this fix has the following changes:

* Remove DISTINCT through FilterApi for list views
* Maintain same functionality using php to filter distinct records from the result set
* Compare the raw SQL result set with the filtered one and compensate the next_offset based on the difference