BR-1899: Composer failing on unit tests
By using phpunit through composer it implicitly includes composer's autloader. However SugarCRM relies on its own autoloader functionality - and we want to keep it this way. The unit tests which was failing was specifically designed to test Sugar's autoloader and an optional $sugar_config parameter to be able to include Composer's autoloader. This combination is not ideal in the sense that phpunit is running in another autoloader environment, regardless of what is set in $sugar_config.

This fix addressses the following:
* Remove $sugar_config option to be able to include composer's autoloader directly
* Solely rely on SugarAutoLoader without the need to include composer's autoloader
* Consume the generated mappings from composer directly within SugarAutoLoader (include_paths, psr0/4 and classmap)
* Make sure to prepend SugarAutoLoader on the top of the spl_autoload stack
* Introduced PSR4 namespace support in SugarAutoLoader
* Add support for multi path definitions for PSR0/4
* Added developer mode support to ignore filemap/classmap for better developer support without fully rebuilding the filemap on every page call