BR-1948: Make SugarAutoLoader and composer integration happy on Windows platform
This fix enforces '/' as directory separator within the SugarAutoLoader functionality and on the SUGAR_BASE_DIR constant. This normalization also applies when consuming the mappings generated by Composer as they may contain Windows directory separators. Once we are in the SugarAutoLoader internals every path is guaranteed to make use of '/' separators. This also implies only having '/' in the file_map and class_map being generated in the cache directory. Added unit test to cover path normalization when run on different platforms.