BR-2022: Refactor earlier fix
Remove the 'installing' flag and ability to reset SugarConfig unique_key value. This refactor handles the setup of unique_key in a proper way for the SilentInstaller (where the actual problem is) and also refactors the handling of options setup_fts_* parameters in a central way before instantiating the SugarSearchEngine object.