PAT-954 1. Use deleted = 0 filter, because we only care about non-deleted records, that is, if there are zero non-deleted records (i.e. non-zero deleted records) 2. Remove needless config param.
Why was this config param added I don't know. I couldn't find history on Git. These are the only 2 places that are using it.
If we need to keep this config param, then another solution would be to show the "Clean up unused combinations of teams" only when the param is enabled.