PAT-1050 `asDb` and `asDbType` (converted to DB time, GMT) now will set the correct timezone. E.g. I create a Task at 19:39 (GMT+2) while my personal tz is set to GMT+7, but consider server time is GMT+2. Record in DB is stored as GMT (2014-10-16 17:39:27).
By using the Gist I provided:

TimeDate->parseDateRange()->parseFromTeplate() will adjust the timezone for the current user, thus giving:
$dates = 0 => 2014-10-17 00:00:00, 1 => 2014-10-17 23:59:59

Now the code goes to $where->lte and $where->gte within the else branch. Calling it with true, will set the dates to GMT i.e.
2014-10-17 14:59:59 and 2014-10-16 15:00:00

(If we call it with false, we get 2014-10-17 23:59:59 and 2014-10-17 00:00:00)

In this range will belong the Task created.