BR-2039: Reuse already existing join for relationship table alias
Although this logic resides in a SugarRelationship class, for the join relationship table alias we may not set $isLink when calling SugarQuery::getJoinTableAlias to be able to reuse already existing joins for that table. See SugarQuery::getJoinTableAlias and SugarQuery::getJoinAlias how the alias is generated. This fix also adds a unit test for SugarQuery::joinSubpanel.