jira: MAR-2440 A JavaScript error occurred -- a circular reference -- when attempting to JSON stringify the model to send the email.
Tooltips add data attributes -- including one for `bs.tooltip` -- to the element to which they are attached. The email options were being pulled from the data attributes on the button. These options were used to create a model, where `bs.tooltip` became one of the attributes. The `bs.tooltip` attribute either contained a circular reference or it created a circular reference with the model. No options were actively being used for launching the email client from the button, so forcing the options to be empty eliminated the circular reference.