jira: MAR-2445 The repeat count default and validation should be UI constructs only.
Default values, column constraints, and validators can only be assigned to individual fields in a vardef. There is not currently a mechanism for validating one field based the values from other fields. The repeat count for an event therefore cannot have a default of 10 since the default will go into the database even when the repeat until value is representative of the true state of the event series. These rules had to be ported to the client, which is the same place that they existing prior to 7.6.