[DNB-26] Megamenu support for fields
This commit adds support for using fields in a mega menu. Once type is added to the metadata of an action in the header, it will treat that action as a field instead of a link. If no type is specified, it will be considered a link.

Please see the sample below for an example usage. The first item and third item in the Array are links. The second item in the array is a field of label type.

Sample:

$moduleName = 'Accounts';
$viewdefs[$moduleName]['base']['menu']['header'] = array(
    array(
        'label' =>'LNK_NEW_ACCOUNT',
        'acl_action'=>'create',
        'acl_module'=>$moduleName,
        'icon' => 'fa-plus',
        'route'=>'#'.$moduleName.'/create',
    ),
    array(
        'name' => 'title',
        'type' => 'label',
        'default_value' => 'LBL_BAL',
    ),
    array(
        'route'=>'#'.$moduleName,
        'label' =>'LNK_ACCOUNT_LIST',
        'acl_action'=>'list',
        'acl_module'=>$moduleName,
        'icon' => 'fa-bars',
    ),