SC-3883 [Merge Duplicates] Sidepane changes Three functional changes were completed. 1) Merge Records RHS pane collapsed by default 2) Merge Records RHS pane setting stored separately from parent module 3) Merge Records RHS pane setting sticky on a per module basis
One code change was also done, which modifies the way the side pane
state is stored in local storage.

Previously, if the hide key existed, then the RHS pane would be hidden,
and if the key didn't exist, it would be shown.

Now, if the key exists and is a "0", it is hidden.
If it is a "1" it is shown.
If the key doesn't exist, then the default behavior (defined by
the instance of default.js) will trigger.