jira: MAR-2441 Unsetting the attribute when the default value is an empty string is not ideal because an empty string is a valid value.
If the value is invalid, then we want the value to be updated the next time the model is synchronized with the server. If the attribute is unset, then it may not be accurately reflected in the database when synchronizing because the API may not handle an unset value appropriately.