Feedback view needs to be re-rendered on toggle
Due to the bad design of the popover plugin, each time the popup is hidden, it
will dispose all the events (click events) that we were listening to.
For that, on toggle we need to force the render of the view again and apply the
plugin for each `toggle(true)`.