BR-2645: Add ability to register same REST end point for multiple request types.
For example when accepting a GET request with a request body containing parameters to avoid (too) long URL's, we also want to register the same entry point as a POST to facility the usage of this entry point by clients which do not support request bodies on GET requests. Sidecar supports this, but for example POSTMAN does not out of the box. The POST entry point is merely meant to cover such clients althought its encouraged to used the correct verb. I.e. hitting /globalsearch implies a GET.