BR-2364: Address random unit test failure
The session clear needs to happen before running the test. Logic hooks triggered *_save when the roles are created can potentially reinitialize the ACL cache already which gives false results on the assertions in this test. For example the after_save logic hook triggered when adding the default roles calls HookHandler of the new SearchEngine framework. The metadatahelper compiles a list of available modules for current user which implicitely refreshes ACL caches. Because this is done before all fixtures are setup in this test, we need to clear the ACL session *after* we are done setting up the fixtures and not before.