Refactored SMTPProxy to be a real proxy
We want to test if SMTP::smtp_conn is a resource in SMTP::connected() per SI Bug #54111. Prior to PHPMailer v5.2.9, SMTP::get_lines() would test for a resource, but it was too late because SMTP::sendCommand() called SMTP::client_send() before SMTP::get_lines(). SMTP::client_send() would call fwrite, which expected a resource. As of PHPMailer v5.2.9, SMTP::sendCommand() calls SMTP::connected() prior to calling SMTP::client_send() and SMTP::connected() tests if SMTP::smtp_conn is a resource. This change is what enabled this refactoring, which is ideal because it eliminates any overrides that cannot be handled through SMTPProxy::__call().